+++ packages/terra-framework-docs/src/terra-dev-site/doc/time-input/example/TimeInputSecondsTwelveHour.jsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import TimeInput from 'terra-time-input/lib/TimeInput';
 import TimeUtil from 'terra-time-input/lib/TimeUtil';
-import classNames from 'classnames/bind';
-import styles from './TimeInputDocCommon.module.scss';
-
-const cx = classNames.bind(styles);
+import Field from 'terra-form-field';
 
 class timeInput extends React.Component {
   constructor(props) {
@@ -20,17 +17,19 @@ class timeInput extends React.Component {
   render() {
     return (
       <div>
-        <p>
-          Time Provided:
-          <span className={cx('time-wrapper')}>{this.state.time}</span>
-        </p>
-        <TimeInput
-          name="time-input-value"
-          value={this.state.time}
-          onChange={this.handleTimeChange}
-          showSeconds
-          variant={TimeUtil.FORMAT_12_HOUR}
-        />
+        <Field
+          label={`Enter Time: ${this.state.time}`}
+          htmlFor="time-input-value"
+        >
+          <TimeInput
+            a11yLabel="Enter Time"
+            name="time-input-value"
+            value={this.state.time}
+            onChange={this.handleTimeChange}
+            showSeconds
+            variant={TimeUtil.FORMAT_12_HOUR}
+          />
+        </Field>
       </div>
     );
   }