+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.54.0 - (January 2, 2024)
+
 * Changed
   * Updated `terra-time-input` example to have visual label.
   * Updated `terra-embedded-content-consumer` example to fix axe error.